requirements.txt
Django==3.0.5
django-cors-headers==3.4.0
djangorestframework==3.11.0
limesurveyrc2api==1.0.0
dj-database-url==0.5.0
razorpay==1.2.0
requests==2.24.0
python-decouple==3.3
PyPDF2==1.26.0
whitenoise==5.1.0
six==1.15.0
Unipath==1.1
python-dateutil==2.8.1
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
idna==2.10
# kiwisolver==1.2.0
pdfminer3k==1.3.4
ply==3.11
pyparsing==2.4.7
pytz==2020.1
slate3k==0.5.3
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
# matplotlib==3.2.2
# numpy==1.19.0